# Faster SCDCalibratePanels