--- a/Framework/DataObjects/inc/MantidDataObjects/Peak.h
+++ b/Framework/DataObjects/inc/MantidDataObjects/Peak.h
@@ -48,7 +48,7 @@ public:
 // MSVC 2015 won't build with noexcept.
 // error C2610: 'Mantid::DataObjects::Peak::Peak(Mantid::DataObjects::Peak &&)
 // noexcept': is not a special member function which can be defaulted
-#if defined(_MSC_VER) && _MSC_VER <= 1700
+#if defined(_MSC_VER) && _MSC_VER <= 1900
   Peak(Peak &&);
   Peak &operator=(Peak &&);
 #else