--- a/Framework/DataObjects/inc/MantidDataObjects/PeaksWorkspace.h
+++ b/Framework/DataObjects/inc/MantidDataObjects/PeaksWorkspace.h
@@ -107,14 +107,14 @@ public:
   const Peak &getPeak(int peakNum) const override;
 
   Geometry::IPeak *createPeak(
-      Kernel::V3D QLabFrame,
+      const Kernel::V3D &QLabFrame,
       boost::optional<double> detectorDistance = boost::none) const override;
   std::vector<std::pair<std::string, std::string>>
-  peakInfo(Kernel::V3D qFrame, bool labCoords) const override;
+  peakInfo(const Kernel::V3D &qFrame, bool labCoords) const override;
 
-  Peak *createPeakHKL(Kernel::V3D HKL) const override;
+  Peak *createPeakHKL(const Kernel::V3D &HKL) const override;
 
-  int peakInfoNumber(Kernel::V3D qFrame, bool labCoords) const override;
+  int peakInfoNumber(const Kernel::V3D &qFrame, bool labCoords) const override;
 
   std::vector<Peak> &getPeaks();
   const std::vector<Peak> &getPeaks() const;