… Framework/Crystal/test/IndexSXPeaksTest.h
… Framework/DataObjects/src/PeaksWorkspace.cpp
… Framework/Geometry/inc/MantidGeometry/Crystal/IPeak.h
… Framework/Geometry/inc/MantidGeometry/Instrument.h
… Framework/Geometry/src/Instrument.cpp
… Framework/Geometry/src/Instrument/ParameterMap.cpp
… Framework/Geometry/test/MockObjects.h
… Framework/Kernel/inc/MantidKernel/Matrix.h
… Framework/Kernel/src/Matrix.cpp
… Framework/PythonInterface/mantid/api/src/Exports/IPeak.cpp
… Framework/PythonInterface/mantid/api/src/Exports/IPeaksWorkspace.cpp
… MantidQt/SliceViewer/test/MockObjects.h