Description of work.

Much of the time in `SCDCalibratePanels` is spent running `PeaksWorkspace::removePeak(int )`. This can be done more efficiently with `std::remove_if` 
followed by `std::vector::erase`. Adding move constructors and move assignment operators for `DataObjects::Peak` makes this step even faster. Passing parameters by const reference also helps. 

On my mac, the test script which was taking 4 minutes, 22 seconds now takes 2 minute 29 seconds. On ndav3 with `OMP_NUM_THREADS=1`, the same script goes from taking 2 minutes 38 seconds to 1 minute 56 seconds.

**To test:**

<!-- Instructions for testing. -->

Build branch `16199_SCD_calib_issues` and time the test script suggested in #17153.
Build branch `16199_SCD_calib_issues` and time the same test script.
**To test:**

<!-- Instructions for testing. -->

There is no GitHub issue associated with this pull request

*Does not need to be in the release notes.*

---

#### Reviewer ####

Please comment on the following ([full description](http://www.mantidproject.org/Individual_Ticket_Testing)):

##### Code Review #####

- [ ] Is the code of an acceptable quality?
- [ ] Does the code conform to the [coding standards](http://www.mantidproject.org/Coding_Standards)? Is it well structured with small focussed classes/methods/functions?
- [ ] Are there unit/system tests in place? Are the unit tests small and test the a class in isolation?
- [ ] If there are changes in the release notes then do they describe the changes appropriately?

##### Functional Tests #####

- [ ] Do changes function as described? Add comments below that describe the tests performed?
- [ ] How do the changes handle unexpected situations, e.g. bad input?
- [ ] Has the relevant documentation been added/updated?
- [ ] Is user-facing documentation written in a user-friendly manner?
- [ ] Has developer documentation been updated if required?

- Does everything look good? Comment with the ship it emoji but don't merge. A member of `@mantidproject/gatekeepers` will take care of it.