@VickieLynch 2016-08-22
On my Rhel7 machine: 2m 16s to 1m 25s for calibration in old pull request and answers are the same.  I did use ZOOM for profiiling and saw that deleting peaks was taking the most time.  Did you try copying the peaks from each bank into a new peaks workspace to see if that is faster than deleting?
:shipit when tests finish.  Great work Steven!