@quantumsteve 2016-08-23
@VickieLynch No, I didn't try copying or moving peaks into a new peaks workspace. I think it would be faster than than repeatedly calling `std::vector::erase(...)`, but don't see why it would be faster than the erase-remove idiom.